Code review notes
