+++ front/backend/apollo-client.js
@@ -5,6 +5,4 @@ const client = new ApolloClient({
   cache: new InMemoryCache(),
 });
 
-console.log(process.env.API_ADDRESS);
-
 export default client;
